Move effects duration parser to shared.js
