+++ js/directives/shared.js
@@ -32,3 +32,45 @@ export function toggleBooleanStateDirective(el, directive, isTruthy, cachedDispl
         el.style.display = isTruthy ? display : 'none'
     }
 }
+
+function parseEffectsDuration(component, expression, key)
+{
+    let target = component
+    let name = expression
+
+    // 1) wire:model.live="$parent.foo" commits on the parent — look up parent effects
+    if (expression.startsWith('$parent')) {
+        target = component.parent
+        name = expression.replace(/^\$parent\.?/, '')
+    }
+
+    let parenthesesIndex = name.indexOf('(')
+    if (parenthesesIndex !== -1) {
+        name = name.slice(0, parenthesesIndex).trim()
+    }
+
+    // 2) Mount-time effect — prefer originalEffects because
+    //    mergeNewSnapshot replaces component.effects on subsequent requests
+    //    (same durability pattern as originalEffects.url / scripts)
+    let options = (target.originalEffects && target.originalEffects[key])
+        || target.effects[key]
+        || {}
+
+    return options[name]
+}
+
+export function debounceEffectDuration(component, expression)
+{
+    let duration = parseEffectsDuration(component, expression, 'debounce')
+
+    // If attribute duration omitted let it fall to default duration
+    // properties: 150ms, methods: 250ms
+    if (duration === true) return undefined
+
+    return ! isNaN(duration) ? duration : undefined
+}
+
+export function hasDebounceEffect(component, expression)
+{
+    return parseEffectsDuration(component, expression, 'debounce') !== undefined
+}
